made UserTest and LocationTest lightweight
