+++ api/src/test/java/org/openmrs/LocationContextSensitiveTest.java
@@ -1,0 +1,34 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs;
+
+import org.junit.jupiter.api.Test;
+import org.openmrs.api.context.Context;
+import org.openmrs.test.jupiter.BaseContextSensitiveTest;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class LocationContextSensitiveTest extends BaseContextSensitiveTest {
+
+	/**
+	 * @see Location#serialize()
+	 * @see Location#hydrate(String)
+	 */
+	@Test
+	public void serializeAndHydrate_shouldRoundTripLocation() {
+		Location location = new Location();
+		location.setName("Test Location");
+		Location originalLocation = Context.getLocationService().saveLocation(location);
+		String serialized = originalLocation.serialize();
+		Location hydratedLocation = originalLocation.hydrate(serialized);
+		assertEquals(originalLocation.getLocationId(), hydratedLocation.getLocationId());
+	}
+
+}
