+++ api/src/test/java/org/openmrs/LocationTest.java
@@ -10,13 +10,10 @@
 package org.openmrs;
 
 import org.junit.jupiter.api.Test;
-import org.openmrs.api.context.Context;
-import org.openmrs.test.jupiter.BaseContextSensitiveTest;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class LocationTest extends BaseContextSensitiveTest {
+public class LocationTest {
 
 	/**
 	 * Get locations that have any of specified set of tags among its child tags.
@@ -36,17 +33,4 @@ public class LocationTest extends BaseContextSensitiveTest {
 		assertTrue(Location.isInHierarchy(locationChild, locationGrandParent));
 	}
 
-	/**
-	 * @see Location#serialize()
-	 * @see Location#hydrate(String)
-	 */
-	@Test
-	public void serializeAndHydrate_shouldRoundTripLocation() {
-		Location location = new Location();
-		location.setName("Test Location");
-		Location originalLocation = Context.getLocationService().saveLocation(location);
-		String serialized = originalLocation.serialize();
-		Location hydratedLocation = originalLocation.hydrate(serialized);
-		assertEquals(originalLocation.getLocationId(), hydratedLocation.getLocationId());
-	}
 }
