+++ api/src/test/java/org/openmrs/UserContextSensitiveTest.java
@@ -1,0 +1,32 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs;
+
+import org.junit.jupiter.api.Test;
+import org.openmrs.api.context.Context;
+import org.openmrs.test.jupiter.BaseContextSensitiveTest;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class UserContextSensitiveTest extends BaseContextSensitiveTest {
+
+	/**
+	 * @see User#serialize()
+	 * @see User#hydrate(String)
+	 */
+	@Test
+	public void serializeAndHydrate_shouldRoundTripUser() {
+		User originalUser = Context.getUserService().getUser(1);
+		String serialized = originalUser.serialize();
+		User hydratedUser = originalUser.hydrate(serialized);
+		assertEquals(originalUser.getUserId(), hydratedUser.getUserId());
+	}
+
+}
